Skew Normal Distribution Parameter Estimation

Function to fit the parameters of the skew normal distribution.

snormFit(x, ...)

Arguments

x

a numeric vector of quantiles.

...

parameters parsed to the optimization function nlm.

Value

snormFit returns a list with the following components:

par

The best set of parameters found.

objective

The value of objective corresponding to par.

convergence

An integer code. 0 indicates successful convergence.

message

A character string giving any additional information returned by the optimizer, or NULL. For details, see PORT documentation.

iterations

Number of iterations performed.

evaluations

Number of objective function and gradient function evaluations.

References

Fernandez C., Steel M.F.J. (2000); On Bayesian Modelling of Fat Tails and Skewness, Preprint, 31 pages.

Author

Diethelm Wuertz for the Rmetrics R-port.

Examples

## rsnorm -
   set.seed(1953)
   r = rsnorm(n = 1000)
       
## snormFit -
   snormFit(r)
#> $par
#>       mean         sd         xi 
#> 0.06694004 1.03494430 1.57611278 
#> 
#> $objective
#> [1] 1413.746
#> 
#> $convergence
#> [1] 0
#> 
#> $iterations
#> [1] 24
#> 
#> $evaluations
#> function gradient 
#>       31       90 
#> 
#> $message
#> [1] "relative convergence (4)"
#>