Skew Student-t Distribution Parameter Estimation

Function to fit the parameters of the Student-t distribution.

sstdFit(x, ...)

Arguments

x

a numeric vector of quantiles.

...

parameters parsed to the optimization function nlm.

Value

sstdFit returns a list with the following components:

par

The best set of parameters found.

objective

The value of objective corresponding to par.

convergence

An integer code. 0 indicates successful convergence.

message

A character string giving any additional information returned by the optimizer, or NULL. For details, see PORT documentation.

iterations

Number of iterations performed.

evaluations

Number of objective function and gradient function evaluations.

References

Fernandez C., Steel M.F.J. (2000); On Bayesian Modelling of Fat Tails and Skewness, Preprint, 31 pages.

Author

Diethelm Wuertz for the Rmetrics R-port.

Examples

## sstd -
   set.seed(1953)
   r = rsstd(n = 1000)

       
## sstdFit -
   sstdFit(r)
#> $minimum
#> [1] 1306.033
#> 
#> $estimate
#>        mean          sd          nu          xi 
#> -0.01376393  1.00314970  4.21091042  1.37308787 
#> 
#> $gradient
#>          mean            sd            nu            xi 
#>  5.693437e-04  3.685488e-04  3.012995e-05 -1.286657e-04 
#> 
#> $code
#> [1] 1
#> 
#> $iterations
#> [1] 18
#>