Extract GARCH model fitted values

Extracts fitted values from a fitted GARCH object.

Methods

Methods for fitted defined in package fGarch:

object = "fGARCH"

Extractor function for fitted values.

Details

fitted() is a generic function which extracts fitted values from objects returned by modeling functions. methods(fitted) will show the methods available in the current R session.

The method for "fGARCH" objects extracts the @fitted value slot from an object of class "fGARCH" as returned by the function garchFit.

The class of the returned value depends on the input to the function garchFit who created the object. The returned value is always of the same class as the input object to the argument data in the function garchFit, i.e. if you fit a "timeSeries" object, you will get back from the function fitted also a "timeSeries" object, if you fit an object of class "zoo", you will get back again a "zoo" object. The same holds for a "numeric" vector, for a "data.frame", and for objects of class "ts", "mts".

In contrast, the slot itself returns independent of the class of the data input always a numeric vector, i.e. the function call rslot(object, "fitted") will return a numeric vector.

Note

(GNB) Contrary to the description of the returned value of the "fGARCH" method, it is always "numeric".

TODO: either implement the documented behaviour or fix the documentation.

Author

Diethelm Wuertz for the Rmetrics R-port

See also

Examples

stopifnot(require("timeSeries"))
## Swiss Pension fund Index -
   x = as.timeSeries(data(LPP2005REC))

## garchFit -
   # Fit LPP40 Bechmark:
   fit = garchFit(LPP40 ~ garch(1, 1), data = 100*x, trace = FALSE)
   fit
#> 
#> Title:
#>  GARCH Modelling 
#> 
#> Call:
#>  garchFit(formula = LPP40 ~ garch(1, 1), data = 100 * x, trace = FALSE) 
#> 
#> Mean and Variance Equation:
#>  LPP40 ~ garch(1, 1)
#> <environment: 0x55e8597d26f0>
#>  [data = 100 * x]
#> 
#> Conditional Distribution:
#>  norm 
#> 
#> Coefficient(s):
#>        mu      omega     alpha1      beta1  
#> 0.0491045  0.0083931  0.0881843  0.8016409  
#> 
#> Std. Errors:
#>  based on Hessian 
#> 
#> Error Analysis:
#>         Estimate  Std. Error  t value Pr(>|t|)    
#> mu      0.049104    0.013478    3.643 0.000269 ***
#> omega   0.008393    0.003493    2.403 0.016268 *  
#> alpha1  0.088184    0.035316    2.497 0.012524 *  
#> beta1   0.801641    0.071544   11.205  < 2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Log Likelihood:
#>  -38.61265    normalized:  -0.1024208 
#> 
#> Description:
#>  Fri Nov  4 08:40:01 2022 by user: georgi 
#> 

## fitted -
   # Fitted values are now a "timeSeries" oject:
   fitted = fitted(fit)
   head(fitted)
#> 2005-11-01 2005-11-02 2005-11-03 2005-11-04 2005-11-07 2005-11-08 
#>  0.0199980 -0.1120404  0.3317548  0.2421248  0.2246611  0.0962708 
   class(fitted)
#> [1] "numeric"

## slot -
   # The slot contains a numeric Vector:
   fitted = slot(fit, "fitted")
   head(fitted)
#> 2005-11-01 2005-11-02 2005-11-03 2005-11-04 2005-11-07 2005-11-08 
#> 0.04910447 0.04910447 0.04910447 0.04910447 0.04910447 0.04910447 
   class(fitted)
#> [1] "numeric"