Extract GARCH Model formula

Extracts formula from a formula GARCH object.

Methods

object = "ANY"

Generic function.

object = "fGARCH"

Extractor function for formula expression.

Details

The function extracts the @formula expression slot from an object of class "fGARCH" as returned by the function garchFit.

Note, the returned formula has always a left hand side. If the argument data was an univariate time series and no name was specified to the series, then the left hand side has assigned the name of the data.set. In the multivariate case the rectangular data object must always have column names, otherwise the fitting will be stopped and you get the error message

The class of the returned value depends on the input to the function garchFit who created the object. The returned value is always of the same class as the input object to the argument data in the function garchFit, i.e. if you fit a "timeSeries" object, you will get back from the function fitted also a "timeSeries" object, if you fit an object of class "zoo", you will get back again a "zoo" object. The same holds for a "numeric" vector, for a "data.frame", and for objects of class "ts", "mts".

In contrast, the slot itself returns independent of the class of the data input always a numeric vector, i.e. the function call rslot(object, "fitted") will return a numeric vector.

Note

formula is a generic function which extracts the formula expression from objects returned by modeling functions.

Author

Diethelm Wuertz for the Rmetrics R-port.

Examples

## garchFit -
   fit = garchFit(~garch(1, 1), data = garchSim())
#> 
#> Series Initialization:
#>  ARMA Model:                arma
#>  Formula Mean:              ~ arma(0, 0)
#>  GARCH Model:               garch
#>  Formula Variance:          ~ garch(1, 1)
#>  ARMA Order:                0 0
#>  Max ARMA Order:            0
#>  GARCH Order:               1 1
#>  Max GARCH Order:           1
#>  Maximum Order:             1
#>  Conditional Dist:          norm
#>  h.start:                   2
#>  llh.start:                 1
#>  Length of Series:          100
#>  Recursion Init:            mci
#>  Series Scale:              0.004019116
#> 
#> Parameter Initialization:
#>  Initial Parameters:          $params
#>  Limits of Transformations:   $U, $V
#>  Which Parameters are Fixed?  $includes
#>  Parameter Matrix:
#>                      U           V     params includes
#>     mu     -0.51941891   0.5194189 0.05194189     TRUE
#>     omega   0.00000100 100.0000000 0.10000000     TRUE
#>     alpha1  0.00000001   1.0000000 0.10000000     TRUE
#>     gamma1 -0.99999999   1.0000000 0.10000000    FALSE
#>     beta1   0.00000001   1.0000000 0.80000000     TRUE
#>     delta   0.00000000   2.0000000 2.00000000    FALSE
#>     skew    0.10000000  10.0000000 1.00000000    FALSE
#>     shape   1.00000000  10.0000000 4.00000000    FALSE
#>  Index List of Parameters to be Optimized:
#>     mu  omega alpha1  beta1 
#>      1      2      3      5 
#>  Persistence:                  0.9 
#> 
#> 
#> --- START OF TRACE ---
#> Selected Algorithm: nlminb 
#> 
#> R coded nlminb Solver: 
#> 
#>   0:     142.03620: 0.0519419 0.100000 0.100000 0.800000
#>   1:     141.90342: 0.0519423 0.110223 0.0999189 0.808069
#>   2:     141.75467: 0.0519434 0.110138 0.0872191 0.805182
#>   3:     141.64947: 0.0519445 0.119798 0.0821376 0.812288
#>   4:     141.52622: 0.0519462 0.119936 0.0692646 0.810314
#>   5:     141.42448: 0.0519488 0.128775 0.0627261 0.817296
#>   6:     141.34973: 0.0519526 0.129785 0.0497539 0.817868
#>   7:     141.28167: 0.0519612 0.136931 0.0426251 0.826098
#>   8:     141.25319: 0.0519816 0.136779 0.0305898 0.831058
#>   9:     141.24246: 0.0520935 0.137930 0.0239640 0.842002
#>  10:     141.23887: 0.0523451 0.144754 0.0250274 0.832079
#>  11:     141.23807: 0.0529381 0.139711 0.0273473 0.834991
#>  12:     141.23790: 0.0534004 0.133373 0.0264942 0.842028
#>  13:     141.23765: 0.0540616 0.135474 0.0262755 0.840258
#>  14:     141.23747: 0.0554580 0.137565 0.0263346 0.838198
#>  15:     141.23746: 0.0558899 0.137883 0.0264450 0.837746
#>  16:     141.23746: 0.0559358 0.137780 0.0264705 0.837811
#>  17:     141.23746: 0.0559333 0.137744 0.0264726 0.837843
#> 
#> Final Estimate of the Negative LLH:
#>  LLH:  -410.4319    norm LLH:  -4.104319 
#>           mu        omega       alpha1        beta1 
#> 2.248026e-04 2.225011e-06 2.647257e-02 8.378433e-01 
#> 
#> R-optimhess Difference Approximated Hessian Matrix:
#>                  mu         omega        alpha1         beta1
#> mu     -6423837.506  9.272300e+07  3.494482e+03  1.029849e+03
#> omega  92722998.166 -6.747940e+12 -1.032148e+08 -1.081282e+08
#> alpha1     3494.482 -1.032148e+08 -1.944143e+03 -1.682497e+03
#> beta1      1029.849 -1.081282e+08 -1.682497e+03 -1.743082e+03
#> attr(,"time")
#> Time difference of 0.007119417 secs
#> 
#> --- END OF TRACE ---
#> 
#> 
#> Time to Estimate Parameters:
#>  Time difference of 0.02715898 secs
   
## formula - 
   formula(fit)
#> data ~ garch(1, 1)
#> attr(,"data")
#> [1] "data = garchSim()"
#> <environment: 0x556851b84738>

## A Bivariate series and mis-specified formula:
   x = garchSim(n = 500)
   y = garchSim(n = 500)
   z = cbind(x, y)
   colnames(z)
#> [1] "garch.1" "garch.2"
   class(z)
#> [1] "timeSeries"
#> attr(,"package")
#> [1] "timeSeries"
   if (FALSE) {
   garchFit(z ~garch(1, 1), data = z, trace = FALSE)
   }
   # Returns:
   # Error in .garchArgsParser(formula = formula, data = data, trace = FALSE) :  
   #   Formula and data units do not match.
   
## Doubled column names in data set - formula can't fit:
   colnames(z) <- c("x", "x")
   z[1:6,]
#> GMT
#>                        x             x
#> 2021-06-10 -0.0023426733  0.0026281998
#> 2021-06-11  0.0012113489 -0.0004620617
#> 2021-06-12  0.0001135257  0.0043161485
#> 2021-06-13 -0.0042403482 -0.0015096423
#> 2021-06-14 -0.0093567474  0.0013723818
#> 2021-06-15  0.0026138964 -0.0047791439
   if (FALSE) {
   garchFit(x ~garch(1, 1), data = z, trace = FALSE)
   }
   # Again the error will be noticed:
   # Error in garchFit(x ~ garch(1, 1), data = z) : 
   #   Column names of data are not unique.

## Missing column names in data set - formula can't fit:
   z.mat <- as.matrix(z)
   colnames(z.mat) <- NULL
   z.mat[1:6,]
#>                     [,1]          [,2]
#> 2021-06-10 -0.0023426733  0.0026281998
#> 2021-06-11  0.0012113489 -0.0004620617
#> 2021-06-12  0.0001135257  0.0043161485
#> 2021-06-13 -0.0042403482 -0.0015096423
#> 2021-06-14 -0.0093567474  0.0013723818
#> 2021-06-15  0.0026138964 -0.0047791439
   if (FALSE) {
   garchFit(x ~ garch(1, 1), data = z.mat, trace = FALSE)
   }
   # Again the error will be noticed:
   # Error in .garchArgsParser(formula = formula, data = data, trace = FALSE) : 
   #   Formula and data units do not match