Class "fGARCH"

The class 'fGARCH' represents a model of an heteroskedastic time series process.

Objects from the Class

Objects can be created by calls of the function garchFit. This object is a parameter estimate of an empirical GARCH process.

Slots

call:

Object of class "call": the call of the garch function.

formula:

Object of class "formula": a formula object specifying the mean and variance equations.

method:

Object of class "character": a string denoting the optimization method, by default "Max Log-Likelihood Estimation".

data:

Object of class "list": a list with one entry named x, containing the data of the time series to be estimated, the same as given by the input argument series.

fit:

Object of class "list": a list with the results from the parameter estimation. The entries of the list depend on the selected algorithm, see below.

residuals:

Object of class "numeric": a numeric vector with the (raw, unstandardized) residual values.

fitted:

Object of class "numeric": a numeric vector with the fitted values.

h.t:

Object of class "numeric":

a numeric vector with the conditional variances (\(h_t = \sigma_t^\delta\)).

sigma.t:

Object of class "numeric": a numeric vector with the conditional standard deviations.

title:

Object of class "character": a title string.

description:

Object of class "character": a string with a brief description.

Methods

plot

signature(x = "fGARCH", y = "missing"): plots an object of class "fGARCH".

show

signature(object = "fGARCH"): prints an object of class "fGARCH".

summary

signature(object = "fGARCH"): summarizes an object of class "fGARCH".

predict

signature(object = "fGARCH"): forecasts mean and volatility from an object of class "fGARCH".

fitted

signature(object = "fGARCH"): extracts fitted values from an object of class "fGARCH".

residuals

signature(object = "fGARCH"): extracts fresiduals from an object of class "fGARCH".

volatility

signature(object = "fGARCH"): extracts conditional volatility from an object of class "fGARCH".

coef

signature(object = "fGARCH"): extracts fitted coefficients from an object of class "fGARCH".

formula

signature(x = "fGARCH"): extracts formula expression from an object of class "fGARCH".

Author

Diethelm Wuertz and Rmetrics Core Team

See also

Examples

## simulate a time series, fit a GARCH(1,1) model, and show it:
x <- garchSim( garchSpec(), n = 500)
fit <- garchFit(~ garch(1, 1), data = x, trace = FALSE)
fit # == print(fit) and also == show(fit)
#> 
#> Title:
#>  GARCH Modelling 
#> 
#> Call:
#>  garchFit(formula = ~garch(1, 1), data = x, trace = FALSE) 
#> 
#> Mean and Variance Equation:
#>  data ~ garch(1, 1)
#> <environment: 0x579d327acd90>
#>  [data = x]
#> 
#> Conditional Distribution:
#>  norm 
#> 
#> Coefficient(s):
#>          mu        omega       alpha1        beta1  
#> -8.5097e-05   5.9309e-06   1.3675e-01   2.1822e-01  
#> 
#> Std. Errors:
#>  based on Hessian 
#> 
#> Error Analysis:
#>          Estimate  Std. Error  t value Pr(>|t|)  
#> mu     -8.510e-05   1.320e-04   -0.645   0.5190  
#> omega   5.931e-06   2.393e-06    2.479   0.0132 *
#> alpha1  1.368e-01   6.074e-02    2.251   0.0244 *
#> beta1   2.182e-01   2.686e-01    0.812   0.4165  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Log Likelihood:
#>  2193.921    normalized:  4.387842 
#> 
#> Description:
#>  Tue Mar 26 17:45:18 2024 by user: georgi 
#>