Extract GARCH Model Residuals

Extracts residuals from a fitted GARCH object.

# S4 method for fGARCH
residuals(object, standardize = FALSE)

Arguments

object

an object of class "fGARCH" as returned from garchFit().

standardize

a logical, indicating if the residuals should be standardized.

Methods

object = "ANY"

Generic function

object = "fGARCH"

Extractor function for residual from an object of class "fGARCH".

Details

The function extracts the @residuals slot from an object of class "fGARCH" as returned by the function garchFit.

The class of the returned value depends on the input to the function garchFit who created the object. The returned value is always of the same class as the input object to the argument data in the function garchFit, i.e. if you fit a "timeSeries" object, you will get back from the function fitted also a "timeSeries" object, if you fit an object of class "zoo", you will get back again a "zoo" object. The same holds for a "numeric" vector, for a "data.frame", and for objects of class "ts", "mts".

In contrast, the slot itself returns independent of the class of the data input always a numeric vector, i.e. the function call rslot(object, "fitted") will return a numeric vector.

Note

residuals is a generic function which extracts residual values from objects returned by modeling functions.

Author

Diethelm Wuertz for the Rmetrics R-port.

Examples

## Swiss Pension fund Index -
   stopifnot(require("timeSeries"))
   x = as.timeSeries(data(LPP2005REC))

## garchFit
   fit = garchFit(LPP40 ~ garch(1, 1), data = 100*x, trace = FALSE)
   fit
#> 
#> Title:
#>  GARCH Modelling 
#> 
#> Call:
#>  garchFit(formula = LPP40 ~ garch(1, 1), data = 100 * x, trace = FALSE) 
#> 
#> Mean and Variance Equation:
#>  LPP40 ~ garch(1, 1)
#> <environment: 0x55684c514850>
#>  [data = 100 * x]
#> 
#> Conditional Distribution:
#>  norm 
#> 
#> Coefficient(s):
#>        mu      omega     alpha1      beta1  
#> 0.0491045  0.0083931  0.0881843  0.8016409  
#> 
#> Std. Errors:
#>  based on Hessian 
#> 
#> Error Analysis:
#>         Estimate  Std. Error  t value Pr(>|t|)    
#> mu      0.049104    0.013478    3.643 0.000269 ***
#> omega   0.008393    0.003493    2.403 0.016268 *  
#> alpha1  0.088184    0.035316    2.497 0.012524 *  
#> beta1   0.801641    0.071544   11.205  < 2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Log Likelihood:
#>  -38.61265    normalized:  -0.1024208 
#> 
#> Description:
#>  Sun Oct 23 08:02:53 2022 by user: georgi 
#> 

## residuals -
   res = residuals(fit)
   head(res)
#> [1] -0.02910647 -0.16114487  0.28265033  0.19302033  0.17555663  0.04716633
   class(res)
#> [1] "numeric"

## slot -
   identical(res, slot(fit, "residuals"))
#> [1] TRUE