GARCH Summary Methods

Summary methods for GARCH Modelling.

Methods

object = "ANY"

Generic function

object = "fGARCH"

Summary function for objects of class "fGARCH".

How to read a diagnostic summary report?

The first five sections return the title, the call, the mean and variance formula, the conditional distribution and the type of standard errors:

 
        Title:
         GARCH Modelling 
        
        Call:
         garchFit(~ garch(1, 1), data = garchSim(), trace = FALSE) 
        
        Mean and Variance Equation:
         ~arch(0)
        
        Conditional Distribution:
         norm 
        
        Std. Errors:
         based on Hessian
        

The next three sections return the estimated coefficients, and an error analysis including standard errors, t values, and probabilities, as well as the log Likelihood values from optimization:

        Coefficient(s):
                  mu         omega        alpha1         beta1  
        -5.79788e-05   7.93017e-06   1.59456e-01   2.30772e-01  
        
        Error Analysis:
                 Estimate  Std. Error  t value Pr(>|t|)
        mu     -5.798e-05   2.582e-04   -0.225    0.822
        omega   7.930e-06   5.309e-06    1.494    0.135
        alpha1  1.595e-01   1.026e-01    1.554    0.120
        beta1   2.308e-01   4.203e-01    0.549    0.583
        
        Log Likelihood:
         -843.3991    normalized:  -Inf 
        

The next section provides results on standardized residuals tests, including statistic and p values, and on information criterion statistic including AIC, BIC, SIC, and HQIC:

 
        Standardized Residuals Tests:
                                        Statistic p-Value    
         Jarque-Bera Test   R    Chi^2  0.4172129 0.8117146  
         Shapiro-Wilk Test  R    W      0.9957817 0.8566985  
         Ljung-Box Test     R    Q(10)  13.05581  0.2205680  
         Ljung-Box Test     R    Q(15)  14.40879  0.4947788  
         Ljung-Box Test     R    Q(20)  38.15456  0.008478302
         Ljung-Box Test     R^2  Q(10)  7.619134  0.6659837  
         Ljung-Box Test     R^2  Q(15)  13.89721  0.5333388  
         Ljung-Box Test     R^2  Q(20)  15.61716  0.7400728  
         LM Arch Test       R    TR^2   7.049963  0.8542942  
         
        Information Criterion Statistics:
                 AIC      BIC      SIC     HQIC 
            8.473991 8.539957 8.473212 8.500687  
        

Author

Diethelm Wuertz for the Rmetrics R-port.

Examples

## garchSim -
   x = garchSim(n = 200)

## garchFit - 
   fit = garchFit(formula = x ~ garch(1, 1), data = x, trace = FALSE)
   summary(fit)
#> 
#> Title:
#>  GARCH Modelling 
#> 
#> Call:
#>  garchFit(formula = x ~ garch(1, 1), data = x, trace = FALSE) 
#> 
#> Mean and Variance Equation:
#>  data ~ garch(1, 1)
#> <environment: 0x556853dc0440>
#>  [data = x]
#> 
#> Conditional Distribution:
#>  norm 
#> 
#> Coefficient(s):
#>          mu        omega       alpha1        beta1  
#> -1.6576e-04   2.4994e-06   9.7539e-02   6.2623e-01  
#> 
#> Std. Errors:
#>  based on Hessian 
#> 
#> Error Analysis:
#>          Estimate  Std. Error  t value Pr(>|t|)   
#> mu     -1.658e-04   2.125e-04   -0.780  0.43545   
#> omega   2.499e-06   1.889e-06    1.323  0.18586   
#> alpha1  9.754e-02   6.893e-02    1.415  0.15707   
#> beta1   6.262e-01   2.364e-01    2.649  0.00808 **
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Log Likelihood:
#>  878.9484    normalized:  4.394742 
#> 
#> Description:
#>  Sun Oct 23 08:02:54 2022 by user: georgi 
#> 
#> 
#> Standardised Residuals Tests:
#>                                 Statistic p-Value  
#>  Jarque-Bera Test   R    Chi^2  1.891722  0.388345 
#>  Shapiro-Wilk Test  R    W      0.9916195 0.3036708
#>  Ljung-Box Test     R    Q(10)  9.620188  0.4744229
#>  Ljung-Box Test     R    Q(15)  18.66545  0.2292963
#>  Ljung-Box Test     R    Q(20)  21.52999  0.3665491
#>  Ljung-Box Test     R^2  Q(10)  9.695337  0.4676167
#>  Ljung-Box Test     R^2  Q(15)  13.48398  0.564969 
#>  Ljung-Box Test     R^2  Q(20)  20.63219  0.4190582
#>  LM Arch Test       R    TR^2   9.646792  0.6469161
#> 
#> Information Criterion Statistics:
#>       AIC       BIC       SIC      HQIC 
#> -8.749484 -8.683518 -8.750263 -8.722789 
#>