GARCH summary methods

Summary methods for GARCH modelling.

Methods

Methods for summary defined in package fGarch:

object = "fGARCH"

Summary function for objects of class "fGARCH".

How to read a diagnostic summary report?

The first five sections return the title, the call, the mean and variance formula, the conditional distribution and the type of standard errors:

 
        Title:
         GARCH Modelling 
        
        Call:
         garchFit(~ garch(1, 1), data = garchSim(), trace = FALSE) 
        
        Mean and Variance Equation:
         ~arch(0)
        
        Conditional Distribution:
         norm 
        
        Std. Errors:
         based on Hessian
        

The next three sections return the estimated coefficients and an error analysis including standard errors, t values, and probabilities, as well as the log Likelihood values from optimization:

        Coefficient(s):
                  mu         omega        alpha1         beta1  
        -5.79788e-05   7.93017e-06   1.59456e-01   2.30772e-01  
        
        Error Analysis:
                 Estimate  Std. Error  t value Pr(>|t|)
        mu     -5.798e-05   2.582e-04   -0.225    0.822
        omega   7.930e-06   5.309e-06    1.494    0.135
        alpha1  1.595e-01   1.026e-01    1.554    0.120
        beta1   2.308e-01   4.203e-01    0.549    0.583
        
        Log Likelihood:
         -843.3991    normalized:  -Inf 
        

The next section provides results on standardized residuals tests, including statistic and p values, and on information criterion statistic including AIC, BIC, SIC, and HQIC:

 
        Standardized Residuals Tests:
                                        Statistic p-Value    
         Jarque-Bera Test   R    Chi^2  0.4172129 0.8117146  
         Shapiro-Wilk Test  R    W      0.9957817 0.8566985  
         Ljung-Box Test     R    Q(10)  13.05581  0.2205680  
         Ljung-Box Test     R    Q(15)  14.40879  0.4947788  
         Ljung-Box Test     R    Q(20)  38.15456  0.008478302
         Ljung-Box Test     R^2  Q(10)  7.619134  0.6659837  
         Ljung-Box Test     R^2  Q(15)  13.89721  0.5333388  
         Ljung-Box Test     R^2  Q(20)  15.61716  0.7400728  
         LM Arch Test       R    TR^2   7.049963  0.8542942  
         
        Information Criterion Statistics:
                 AIC      BIC      SIC     HQIC 
            8.473991 8.539957 8.473212 8.500687  
        

Author

Diethelm Wuertz for the Rmetrics R-port

Examples

## garchSim -
   x = garchSim(n = 200)

## garchFit - 
   fit = garchFit(formula = x ~ garch(1, 1), data = x, trace = FALSE)
   summary(fit)
#> 
#> Title:
#>  GARCH Modelling 
#> 
#> Call:
#>  garchFit(formula = x ~ garch(1, 1), data = x, trace = FALSE) 
#> 
#> Mean and Variance Equation:
#>  data ~ garch(1, 1)
#> <environment: 0x55e8580751a8>
#>  [data = x]
#> 
#> Conditional Distribution:
#>  norm 
#> 
#> Coefficient(s):
#>         mu       omega      alpha1       beta1  
#> 2.5896e-04  4.1606e-07  1.1732e-01  8.2562e-01  
#> 
#> Std. Errors:
#>  based on Hessian 
#> 
#> Error Analysis:
#>         Estimate  Std. Error  t value Pr(>|t|)    
#> mu     2.590e-04   1.723e-04    1.503   0.1328    
#> omega  4.161e-07   3.378e-07    1.232   0.2181    
#> alpha1 1.173e-01   5.679e-02    2.066   0.0388 *  
#> beta1  8.256e-01   7.644e-02   10.801   <2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Log Likelihood:
#>  904.4709    normalized:  4.522355 
#> 
#> Description:
#>  Fri Nov  4 08:40:02 2022 by user: georgi 
#> 
#> 
#> Standardised Residuals Tests:
#>                                 Statistic p-Value  
#>  Jarque-Bera Test   R    Chi^2  0.2185727 0.8964737
#>  Shapiro-Wilk Test  R    W      0.9979898 0.997206 
#>  Ljung-Box Test     R    Q(10)  2.094741  0.9955611
#>  Ljung-Box Test     R    Q(15)  7.772207  0.93259  
#>  Ljung-Box Test     R    Q(20)  18.3001   0.5676465
#>  Ljung-Box Test     R^2  Q(10)  9.368209  0.4975518
#>  Ljung-Box Test     R^2  Q(15)  15.03517  0.4488868
#>  Ljung-Box Test     R^2  Q(20)  18.64784  0.5448116
#>  LM Arch Test       R    TR^2   13.17633  0.3563554
#> 
#> Information Criterion Statistics:
#>       AIC       BIC       SIC      HQIC 
#> -9.004709 -8.938743 -9.005489 -8.978014 
#>